Exercise 4
Hi, It looks as if I just copy and pasted this in.. because I did. I
forked it wrong the first time and it would let me write the code but
not commit so I saved the code I had already written in word documents
and figured out how to fork it correctly and then pasted my changes
back into the document.